requirements.txt
extcolors==1.0.0
matplotlib==3.6.0
numpy==1.23.3
Pillow==9.2.0
scikit_learn==1.1.2
opencv-python==4.6.0.66
tornado==6.2